@@ -24,25 +24,25 @@
     "build:taro:h5": "npm -C playgrounds/tarojs run build:h5"
   },
   "devDependencies": {
-    "@antfu/eslint-config": "^1.0.0-beta.18",
-    "@iconify-json/tabler": "^1.1.94",
-    "@types/node": "^18.18.4",
+    "@antfu/eslint-config": "^2.13.0",
+    "@iconify-json/tabler": "^1.1.109",
+    "@types/node": "^20.12.5",
     "@unocss-applet/preset-applet": "workspace:*",
     "@unocss-applet/preset-rem-rpx": "workspace:*",
     "@unocss-applet/reset": "workspace:*",
     "@unocss-applet/shared": "workspace:*",
     "@unocss-applet/transformer-attributify": "workspace:*",
-    "@unocss/eslint-config": "^0.56.5",
-    "bumpp": "^9.2.0",
-    "eslint": "^8.51.0",
-    "esno": "^0.17.0",
-    "magic-string": "^0.30.4",
-    "typescript": "^5.2.2",
+    "@unocss/eslint-config": "^0.59.0",
+    "bumpp": "^9.4.0",
+    "eslint": "^8.57.0",
+    "esno": "^4.7.0",
+    "magic-string": "^0.30.9",
+    "typescript": "^5.4.4",
     "unbuild": "^2.0.0",
-    "unocss": "^0.56.5",
+    "unocss": "^0.59.0",
     "unocss-applet": "workspace:*",
-    "unocss-preset-extra": "^0.5.2",
-    "vite": "^4.4.11",
+    "unocss-preset-extra": "^0.5.3",
+    "vite": "^4.5.3",
     "vitest": "^0.34.6"
   }
 }
