@@ -35,37 +35,40 @@
     "ios >= 8"
   ],
   "dependencies": {
-    "@babel/runtime": "^7.23.1",
-    "@tarojs/components": "3.6.17",
-    "@tarojs/helper": "3.6.17",
-    "@tarojs/plugin-framework-vue3": "3.6.17",
-    "@tarojs/plugin-html": "^3.6.17",
-    "@tarojs/plugin-platform-alipay": "3.6.17",
-    "@tarojs/plugin-platform-jd": "3.6.17",
-    "@tarojs/plugin-platform-qq": "3.6.17",
-    "@tarojs/plugin-platform-swan": "3.6.17",
-    "@tarojs/plugin-platform-tt": "3.6.17",
-    "@tarojs/plugin-platform-weapp": "3.6.17",
-    "@tarojs/router": "3.6.17",
-    "@tarojs/runtime": "3.6.17",
-    "@tarojs/shared": "3.6.17",
-    "@tarojs/taro": "3.6.17",
-    "@tarojs/taro-h5": "3.6.17",
-    "vue": "^3.3.4"
+    "@babel/runtime": "^7.24.4",
+    "@tarojs/components": "3.6.25",
+    "@tarojs/helper": "3.6.25",
+    "@tarojs/plugin-framework-vue3": "3.6.25",
+    "@tarojs/plugin-html": "3.6.25",
+    "@tarojs/plugin-platform-alipay": "3.6.25",
+    "@tarojs/plugin-platform-h5": "3.6.25",
+    "@tarojs/plugin-platform-jd": "3.6.25",
+    "@tarojs/plugin-platform-qq": "3.6.25",
+    "@tarojs/plugin-platform-swan": "3.6.25",
+    "@tarojs/plugin-platform-tt": "3.6.25",
+    "@tarojs/plugin-platform-weapp": "3.6.25",
+    "@tarojs/router": "3.6.25",
+    "@tarojs/runtime": "3.6.25",
+    "@tarojs/shared": "3.6.25",
+    "@tarojs/taro": "3.6.25",
+    "@tarojs/taro-h5": "3.6.25",
+    "vue": "^3.4.21"
   },
   "devDependencies": {
-    "@babel/core": "^7.23.0",
-    "@tarojs/cli": "3.6.17",
-    "@tarojs/webpack5-runner": "3.6.17",
-    "@types/webpack-env": "^1.18.2",
-    "@unocss/webpack": "^0.56.5",
-    "@vue/babel-plugin-jsx": "^1.1.5",
-    "@vue/compiler-sfc": "^3.3.4",
-    "babel-preset-taro": "3.6.17",
-    "css-loader": "6.8.1",
-    "style-loader": "3.3.3",
-    "typescript": "^5.2.2",
-    "vue-loader": "^17.3.0",
-    "webpack": "5.88.2"
+    "@babel/core": "^7.24.4",
+    "@tarojs/cli": "3.6.25",
+    "@tarojs/taro-loader": "3.6.25",
+    "@tarojs/webpack5-runner": "3.6.25",
+    "@types/webpack-env": "^1.18.4",
+    "@unocss/webpack": "^0.59.0",
+    "@vue/babel-plugin-jsx": "^1.2.2",
+    "@vue/compiler-sfc": "^3.4.21",
+    "babel-preset-taro": "3.6.25",
+    "css-loader": "^7.1.0",
+    "postcss": "^8.4.18",
+    "style-loader": "^4.0.0",
+    "typescript": "^5.4.4",
+    "vue-loader": "^17.4.2",
+    "webpack": "^5.91.0"
   }
 }
