@@ -33,30 +33,30 @@
     "build:quickapp-webview-union": "uni build -p quickapp-webview-union"
   },
   "dependencies": {
-    "@dcloudio/uni-app": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-app-plus": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-components": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-h5": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-mp-alipay": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-mp-baidu": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-mp-kuaishou": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-mp-lark": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-mp-qq": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-mp-toutiao": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-mp-weixin": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-quickapp-webview": "3.0.0-alpha-3090120230927001",
-    "ano-ui": "^0.8.1",
-    "vue": "^3.3.4",
-    "vue-i18n": "^9.5.0"
+    "@dcloudio/uni-app": "3.0.0-4000820240401001",
+    "@dcloudio/uni-app-plus": "3.0.0-4000820240401001",
+    "@dcloudio/uni-components": "3.0.0-4000820240401001",
+    "@dcloudio/uni-h5": "3.0.0-4000820240401001",
+    "@dcloudio/uni-mp-alipay": "3.0.0-4000820240401001",
+    "@dcloudio/uni-mp-baidu": "3.0.0-4000820240401001",
+    "@dcloudio/uni-mp-kuaishou": "3.0.0-4000820240401001",
+    "@dcloudio/uni-mp-lark": "3.0.0-4000820240401001",
+    "@dcloudio/uni-mp-qq": "3.0.0-4000820240401001",
+    "@dcloudio/uni-mp-toutiao": "3.0.0-4000820240401001",
+    "@dcloudio/uni-mp-weixin": "3.0.0-4000820240401001",
+    "@dcloudio/uni-quickapp-webview": "3.0.0-4000820240401001",
+    "ano-ui": "^0.9.2",
+    "vue": "^3.4.21",
+    "vue-i18n": "^9.11.0"
   },
   "devDependencies": {
-    "@dcloudio/types": "^3.4.0",
-    "@dcloudio/uni-automator": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/uni-cli-shared": "3.0.0-alpha-3090120230927001",
-    "@dcloudio/vite-plugin-uni": "3.0.0-alpha-3090120230927001",
-    "@iconify-json/carbon": "^1.1.21",
-    "postcss": "^8.4.31",
-    "sass": "^1.69.0",
-    "vite-plugin-inspect": "^0.7.40"
+    "@dcloudio/types": "^3.4.8",
+    "@dcloudio/uni-automator": "3.0.0-4000820240401001",
+    "@dcloudio/uni-cli-shared": "3.0.0-4000820240401001",
+    "@dcloudio/vite-plugin-uni": "3.0.0-4000820240401001",
+    "@iconify-json/carbon": "^1.1.31",
+    "postcss": "^8.4.38",
+    "sass": "^1.74.1",
+    "vite-plugin-inspect": "^0.8.3"
   }
 }
