@@ -1,91 +1,95 @@
 import process from 'node:process'
-import UnoCSS from 'unocss/webpack'
+import dev from './dev'
+import prod from './prod'
 
-const config = {
-  projectName: 'tarojs',
-  date: '2022-9-24',
-  designWidth: 750,
-  deviceRatio: {
-    640: 2.34 / 2,
-    750: 1,
-    828: 1.81 / 2,
-  },
-  sourceRoot: 'src',
-  outputRoot: 'dist',
-  plugins: ['@tarojs/plugin-html'],
-  defineConstants: {
-  },
-  copy: {
-    patterns: [
-    ],
-    options: {
-    },
-  },
-  framework: 'vue3',
-  compiler: 'webpack5',
-  cache: {
-    enable: false, // Webpack 持久化缓存配置，建议开启。默认配置请参考：https://docs.taro.zone/docs/config-detail#cache
-  },
-  mini: {
-    postcss: {
-      pxtransform: {
-        enable: true,
-        config: {
+export default async (merge) => {
+  const unocssModule = await import('@unocss/webpack')
+  const UnoCSS = unocssModule.default || unocssModule
 
-        },
-      },
-      url: {
-        enable: true,
-        config: {
-          limit: 1024, // 设定转换尺寸上限
-        },
-      },
-      cssModules: {
-        enable: false, // 默认为 false，如需使用 css modules 功能，则设为 true
-        config: {
-          namingPattern: 'module', // 转换模式，取值为 global/module
-          generateScopedName: '[name]__[local]___[hash:base64:5]',
-        },
+  const base = {
+    projectName: 'tarojs',
+    date: '2022-9-24',
+    designWidth: 750,
+    deviceRatio: {
+      640: 2.34 / 2,
+      750: 1,
+      828: 1.81 / 2,
+    },
+    sourceRoot: 'src',
+    outputRoot: 'dist',
+    plugins: ['@tarojs/plugin-html'],
+    defineConstants: {
+    },
+    copy: {
+      patterns: [
+      ],
+      options: {
       },
     },
-    webpackChain(chain, _webpack) {
-      chain.plugin('unocss').use(UnoCSS())
+    framework: 'vue3',
+    compiler: 'webpack5',
+    cache: {
+      enable: false, // Webpack 持久化缓存配置，建议开启。默认配置请参考：https://docs.taro.zone/docs/config-detail#cache
     },
-  },
-  h5: {
-    publicPath: '/',
-    staticDirectory: 'static',
-    postcss: {
-      autoprefixer: {
-        enable: true,
-        config: {
-        },
-      },
-      cssModules: {
-        enable: false, // 默认为 false，如需使用 css modules 功能，则设为 true
-        config: {
-          namingPattern: 'module', // 转换模式，取值为 global/module
-          generateScopedName: '[name]__[local]___[hash:base64:5]',
-        },
-      },
-    },
-    webpackChain(chain, _webpack) {
-      chain.plugin('unocss').use(UnoCSS())
-    },
-  },
-  rn: {
-    appName: 'taroDemo',
-    postcss: {
-      cssModules: {
-        enable: false, // 默认为 false，如需使用 css modules 功能，则设为 true
-      },
-    },
-  },
-}
+    mini: {
+      postcss: {
+        pxtransform: {
+          enable: true,
+          config: {
 
-module.exports = function (merge) {
+          },
+        },
+        url: {
+          enable: true,
+          config: {
+            limit: 1024, // 设定转换尺寸上限
+          },
+        },
+        cssModules: {
+          enable: false, // 默认为 false，如需使用 css modules 功能，则设为 true
+          config: {
+            namingPattern: 'module', // 转换模式，取值为 global/module
+            generateScopedName: '[name]__[local]___[hash:base64:5]',
+          },
+        },
+      },
+      webpackChain(chain, _webpack) {
+        chain.plugin('unocss').use(UnoCSS())
+      },
+    },
+    h5: {
+      publicPath: '/',
+      staticDirectory: 'static',
+      postcss: {
+        autoprefixer: {
+          enable: true,
+          config: {
+          },
+        },
+        cssModules: {
+          enable: false, // 默认为 false，如需使用 css modules 功能，则设为 true
+          config: {
+            namingPattern: 'module', // 转换模式，取值为 global/module
+            generateScopedName: '[name]__[local]___[hash:base64:5]',
+          },
+        },
+      },
+      webpackChain(chain, _webpack) {
+        chain.plugin('unocss').use(UnoCSS())
+      },
+    },
+    rn: {
+      appName: 'taroDemo',
+      postcss: {
+        cssModules: {
+          enable: false, // 默认为 false，如需使用 css modules 功能，则设为 true
+        },
+      },
+    },
+  }
+
   if (process.env.NODE_ENV === 'development')
-    return merge({}, config, require('./dev'))
+    return merge({}, base, dev)
 
-  return merge({}, config, require('./prod'))
+  return merge({}, base, prod)
 }
